- Changes for treating index page. - Changes between BasePresenter.php and ModuleBasePresenter.php.
